+++ src/router/index.js
@@ -1,45 +1,46 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import Home from '../views/Home.vue'
+
 import BasicExample from '../views/BasicExample.vue'
-import CheckboxExample from '../views/CheckboxExample.vue'
 import AsyncExample from '../views/AsyncExample.vue'
 import EditExample from '../views/EditExample.vue'
-import DragAndDropExample from '../views/DragAndDropExample.vue'
 import ButtonExample from '../views/ButtonExample.vue'
 import ContextMenuExample from '../views/ContextMenuExample.vue'
-import CustomAppearanceExample from '../views/CustomAppearanceExample.vue'
 import SearchExample from '../views/SearchExample.vue'
 import MultiSelectExample from '../views/MultiSelectExample.vue'
 import RadioButtonExample from '../views/RadioButtonExample.vue'
 import SortExample from '../views/SortExample.vue'
 import NoRootExample from '../views/NoRootExample.vue'
 
+import DragAndDropBasicExample from '../views/DragAndDropBasicExample.vue'
+import DragAndDropAdvancedExample from '../views/DragAndDropAdvancedExample.vue'
+import DragAndDropDisableExample from '../views/DragAndDropDisableExample.vue'
+import DragAndDropCustomAppearanceExample from '../views/DragAndDropCustomAppearanceExample.vue'
+
+import CheckboxWithLinkageExample from '../views/CheckboxWithLinkageExample.vue'
+import CheckboxWithoutLinkageExample from '../views/CheckboxWithoutLinkageExample.vue'
+import CheckboxCustomAppearanceExample from '../views/CheckboxCustomAppearanceExample.vue'
+
+import CustomAppearanceSizeExample from '../views/CustomAppearanceSizeExample.vue'
+import CustomAppearanceIconExample from '../views/CustomAppearanceIconExample.vue'
+import CustomAppearanceNoIconsExample from '../views/CustomAppearanceNoIconsExample.vue'
+import CustomAppearanceBackgroundColorExample from '../views/CustomAppearanceBackgroundColorExample.vue'
+import CustomAppearanceStripesExample from '../views/CustomAppearanceStripesExample.vue'
+import CustomAppearanceDisableAnimationsExample from '../views/CustomAppearanceDisableAnimationsExample.vue'
+import CustomAppearanceExtraContentExample from '../views/CustomAppearanceExtraContentExample.vue'
+import CustomAppearanceLimitTitleWidthExample from '../views/CustomAppearanceLimitTitleWidthExample.vue'
+
 Vue.use(VueRouter)
 
 const routes = [
   {
     path: '/',
-    name: 'home',
-    component: Home
-  },
-  {
-    path: '/about',
-    name: 'about',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
+    redirect: '/example/basic',
   },
   {
     path: '/example/basic',
     name: 'basic-example',
     component: BasicExample
-  },
-  {
-    path: '/example/checkbox',
-    name: 'checkbox-example',
-    component: CheckboxExample
   },
   {
     path: '/example/async',
@@ -52,11 +53,6 @@ const routes = [
     component: EditExample
   },
   {
-    path: '/example/drag-and-drop',
-    name: 'drag-and-drop-example',
-    component: DragAndDropExample
-  },
-  {
     path: '/example/button',
     name: 'button-example',
     component: ButtonExample
@@ -66,11 +62,7 @@ const routes = [
     name: 'contextmenu-example',
     component: ContextMenuExample
   },
-  {
-    path: '/example/custom-appearance',
-    name: 'custom-appearance-example',
-    component: CustomAppearanceExample
-  },
+
   {
     path: '/example/search',
     name: 'search-example',
@@ -95,6 +87,85 @@ const routes = [
     path: '/example/no-root',
     name: 'no-root-example',
     component: NoRootExample
+  },
+
+
+  {
+    path: '/example/checkbox/with-linkage',
+    name: 'checkbox-with-linkage-example',
+    component: CheckboxWithLinkageExample
+  },
+  {
+    path: '/example/checkbox/without-linkage',
+    name: 'checkbox-without-linkage-example',
+    component: CheckboxWithoutLinkageExample
+  },
+  {
+    path: '/example/checkbox/custom-appearance',
+    name: 'checkbox-custom-appearance-example',
+    component: CheckboxCustomAppearanceExample
+  },
+
+  {
+    path: '/example/drag-and-drop/basic',
+    name: 'drag-and-drop-basic-example',
+    component: DragAndDropBasicExample
+  },
+  {
+    path: '/example/drag-and-drop/advanced',
+    name: 'drag-and-drop-advanced-example',
+    component: DragAndDropAdvancedExample
+  },
+  {
+    path: '/example/drag-and-drop/disable',
+    name: 'drag-and-drop-disable-example',
+    component: DragAndDropDisableExample
+  },
+  {
+    path: '/example/drag-and-drop/custom-appearance',
+    name: 'drag-and-drop-custom-appearance-example',
+    component: DragAndDropCustomAppearanceExample
+  },
+
+  {
+    path: '/example/custom-appearance/size',
+    name: 'custom-appearance-size-example',
+    component: CustomAppearanceSizeExample
+  },
+  {
+    path: '/example/custom-appearance/icon',
+    name: 'custom-appearance-icon-example',
+    component: CustomAppearanceIconExample
+  },
+  {
+    path: '/example/custom-appearance/no-icons',
+    name: 'custom-appearance-no-icons-example',
+    component: CustomAppearanceNoIconsExample
+  },
+  {
+    path: '/example/custom-appearance/background-color',
+    name: 'custom-appearance-background-color-example',
+    component: CustomAppearanceBackgroundColorExample
+  },
+  {
+    path: '/example/custom-appearance/stripes',
+    name: 'custom-appearance-stripes-example',
+    component: CustomAppearanceStripesExample
+  },
+  {
+    path: '/example/custom-appearance/disable-animations',
+    name: 'custom-appearance-disable-animations-example',
+    component: CustomAppearanceDisableAnimationsExample
+  },
+  {
+    path: '/example/custom-appearance/extra-content',
+    name: 'custom-appearance-extra-content-example',
+    component: CustomAppearanceExtraContentExample
+  },
+  {
+    path: '/example/custom-appearance/limit-title-width',
+    name: 'custom-appearance-limit-title-width-example',
+    component: CustomAppearanceLimitTitleWidthExample
   }
 ]